team34/src/main/webapp/app/entities/entry/list/entry.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpHeaders, HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { EntryService } from '../service/entry.service';

import { EntryComponent } from './entry.component';

describe('Entry Management Component', () => {
  let comp: EntryComponent;
  let fixture: ComponentFixture<EntryComponent>;
  let service: EntryService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [RouterTestingModule.withRoutes([{ path: 'entry', component: EntryComponent }]), HttpClientTestingModule],
      declarations: [EntryComponent],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            data: of({
              defaultSort: 'id,asc',
            }),
            queryParamMap: of(
              jest.requireActual('@angular/router').convertToParamMap({
                page: '1',
                size: '1',
                sort: 'id,desc',
              })
            ),
            snapshot: { queryParams: {} },
          },
        },
      ],
    })
      .overrideTemplate(EntryComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(EntryComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(EntryService);

    const headers = new HttpHeaders();
    jest.spyOn(service, 'query').mockReturnValue(
      of(
        new HttpResponse({
          body: [{ id: 123 }],
          headers,
        })
      )
    );
  });

  it('Should call load all on init', () => {
    // WHEN
    comp.ngOnInit();

    // THEN
    expect(service.query).toHaveBeenCalled();
    expect(comp.entries?.[0]).toEqual(expect.objectContaining({ id: 123 }));
  });

  describe('trackId', () => {
    it('Should forward to entryService', () => {
      const entity = { id: 123 };
      jest.spyOn(service, 'getEntryIdentifier');
      const id = comp.trackId(0, entity);
      expect(service.getEntryIdentifier).toHaveBeenCalledWith(entity);
      expect(id).toBe(entity.id);
    });
  });
});
